third_party/blink/web_tests/platform/generic/fast/events/keydown-numlock-standard-location-expected.txt

Checks that KeyboardEvent.location is DOM_KEY_LOCATION_STANDARD for NumLock key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is VKEY_NUMLOCK
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_STANDARD
PASS successfullyParsed is true

TEST COMPLETE
